+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/invitecode/InviteCodeViewModel.kt
@@ -8,6 +8,7 @@ import com.mulberry.ody.domain.repository.ody.MeetingRepository
 import com.mulberry.ody.presentation.common.BaseViewModel
 import com.mulberry.ody.presentation.common.analytics.AnalyticsHelper
 import com.mulberry.ody.presentation.common.analytics.logNetworkErrorEvent
+import com.mulberry.ody.presentation.feature.invitecode.model.InviteCodeNavigateAction
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
