+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/invitecode/model/InviteCodeNavigateAction.kt
@@ -1,4 +1,4 @@
-package com.mulberry.ody.presentation.feature.invitecode
+package com.mulberry.ody.presentation.feature.invitecode.model
 
 sealed interface InviteCodeNavigateAction {
     data object CodeNavigateToJoin : InviteCodeNavigateAction
