+++ android/app/src/test/java/com/mulberry/ody/presentation/feature/invitecode/InviteCodeViewModelTest.kt
@@ -2,6 +2,7 @@ package com.mulberry.ody.presentation.feature.invitecode
 
 import com.mulberry.ody.fake.FakeAnalyticsHelper
 import com.mulberry.ody.fake.FakeMeetingRepository
+import com.mulberry.ody.presentation.feature.invitecode.model.InviteCodeNavigateAction
 import com.mulberry.ody.util.CoroutinesTestExtension
 import com.mulberry.ody.util.InstantTaskExecutorExtension
 import com.mulberry.ody.util.valueOnAction
